refactor: move JSP tag transformation into substituteTags()
It really had no business being in a function called `extractJS()`.
