+++ packages/liferay-npm-scripts/test/format/substituteTags.js
@@ -5,6 +5,7 @@
  */
 
 const extractJS = require('../../src/format/extractJS');
+const dedent = require('../../support/dedent');
 const substituteTags = require('../../src/format/substituteTags');
 const getFixture = require('../../support/getFixture');
 
@@ -21,11 +22,35 @@ async function getScript(fixture) {
 		);
 	}
 
-	return blocks[0];
+	return blocks[0].contents;
 }
 
 describe('substituteTags()', () => {
 	it('turns Expression Language syntax (${}) into identifier placeholders', () => {});
+
+	it('turns JSP expressions (<%= ... %>) into identifier placeholders', () => {
+		const transformed = substituteTags(dedent(3)`
+			function create() {
+				A.Node.create(
+					'<div class="alert"><%= SomeUtil("abc") %></div>'
+				);
+			}
+		`);
+
+		expect(transformed).toEqual(dedent(3)`
+			function create() {
+				A.Node.create(
+					'<div class="alert">_ECHO_SCRIPTLET_</div>'
+				);
+			}
+		`);
+	});
+
+	it('turns JSP directives (<%@ ... %>) into identifier placeholders', () => {});
+
+	it('turns JSP scriplets (<% ... %>) into comments', () => {
+		// TODO deal with c:if etc, which would ideally produce `if` blocks etc
+	});
 
 	it('turns childless JSP tags into comments', async () => {
 		// See the </liferay-portlet:renderURL> tag in this fixture, which is in
